Refactor smart automation into dedicated server and UI modules
