+++ src/app.module.ts
@@ -17,6 +17,7 @@ import { PlatformToolsModule } from './modules/platform-tools/platform-tools.mod
 import { PluginsModule } from './modules/plugins/plugins.module.js'
 import { ServerModule } from './modules/server/server.module.js'
 import { SetupWizardModule } from './modules/setup-wizard/setup-wizard.module.js'
+import { SmartAutomationsModule } from './modules/smart-automations/smart-automations.module.js'
 import { StatusModule } from './modules/status/status.module.js'
 import { UpdateAllModule } from './modules/update-all/update-all.module.js'
 import { UsersModule } from './modules/users/users.module.js'
@@ -33,6 +34,7 @@ import { UsersModule } from './modules/users/users.module.js'
     UsersModule,
     StatusModule,
     AccessoriesModule,
+    SmartAutomationsModule,
     ConfigEditorModule,
     PlatformToolsModule,
     ChildBridgesModule,
