+++ src/modules/accessories/accessories.gateway.ts
@@ -39,31 +39,4 @@ export class AccessoriesGateway {
       return new WsException(e)
     }
   }
-
-  @SubscribeMessage('get-smart-automations')
-  async getSmartAutomations(client: any, payload: any) {
-    try {
-      return await this.accessoriesService.getSmartAutomations(payload.user)
-    } catch (e) {
-      return new WsException(e)
-    }
-  }
-
-  @SubscribeMessage('save-smart-automation')
-  async saveSmartAutomation(client: any, payload: any) {
-    try {
-      return await this.accessoriesService.saveSmartAutomation(payload.user, payload.automation)
-    } catch (e) {
-      return new WsException(e)
-    }
-  }
-
-  @SubscribeMessage('delete-smart-automation')
-  async deleteSmartAutomation(client: any, payload: any) {
-    try {
-      return await this.accessoriesService.deleteSmartAutomation(payload.user, payload.id)
-    } catch (e) {
-      return new WsException(e)
-    }
-  }
 }
