+++ src/modules/accessories/accessories.service.ts
@@ -3,7 +3,6 @@ import type { Socket } from 'socket.io'
 
 import type { AccessoryControlMessage } from './accessories.interfaces.js'
 
-import { randomUUID } from 'node:crypto'
 import { join } from 'node:path'
 
 import { HapClient } from '@homebridge/hap-client'
@@ -25,15 +24,6 @@ import {
   MatterService,
   MatterStateUpdate,
 } from '../../core/matter/matter.interfaces.js'
-
-export interface SmartAutomation {
-  id: string
-  name: string
-  type: 'smart-light-group'
-  uniqueIds: string[]
-  restoreAfterMs: number
-  enabled: boolean
-}
 
 @Injectable()
 export class AccessoriesService {
@@ -628,88 +618,6 @@ export class AccessoriesService {
     )
     this.logger.log(`Accessory layout changes saved for ${user}.`)
     return layout
-  }
-
-  public async getSmartAutomations(username: string) {
-    try {
-      const smartAutomations = await readJson(this.getSmartAutomationsPath()) as Record<string, SmartAutomation[]>
-      return (smartAutomations[username] || []).map(automation => ({
-        ...automation,
-        enabled: automation.enabled ?? true,
-      }))
-    } catch (e) {
-      return []
-    }
-  }
-
-  public async saveSmartAutomation(username: string, automation: Partial<SmartAutomation>) {
-    if (!automation?.name || typeof automation.name !== 'string') {
-      throw new BadRequestException('Automation name is required.')
-    }
-
-    if (automation.type !== 'smart-light-group') {
-      throw new BadRequestException('Unsupported automation type.')
-    }
-
-    const uniqueIds = [...new Set(automation.uniqueIds || [])].filter(x => typeof x === 'string' && x.length > 0)
-    if (!uniqueIds.length) {
-      throw new BadRequestException('At least one light uniqueId is required.')
-    }
-
-    const restoreAfterMs = Number.isInteger(automation.restoreAfterMs) ? automation.restoreAfterMs! : 30000
-    if (restoreAfterMs < 1000 || restoreAfterMs > 86_400_000) {
-      throw new BadRequestException('restoreAfterMs must be between 1000 and 86400000.')
-    }
-    const enabled = typeof automation.enabled === 'boolean' ? automation.enabled : true
-
-    if (!await pathExists(join(this.configService.storagePath, 'accessories'))) {
-      await mkdirp(join(this.configService.storagePath, 'accessories'))
-    }
-
-    const saved = {
-      id: automation.id || randomUUID(),
-      name: automation.name.trim(),
-      type: 'smart-light-group' as const,
-      uniqueIds,
-      restoreAfterMs,
-      enabled,
-    }
-
-    await this.jsonStore.mutate<Record<string, SmartAutomation[]>>(this.getSmartAutomationsPath(), (current) => {
-      const next = current || {}
-      const userAutomations = next[username] || []
-      const existingIndex = userAutomations.findIndex(x => x.id === saved.id)
-      if (existingIndex === -1) {
-        next[username] = [...userAutomations, saved]
-      } else {
-        next[username] = userAutomations.map(item => item.id === saved.id ? saved : item)
-      }
-      return next
-    })
-
-    return saved
-  }
-
-  public async deleteSmartAutomation(username: string, id: string) {
-    if (!id || typeof id !== 'string') {
-      throw new BadRequestException('Automation id is required.')
-    }
-
-    await this.jsonStore.mutate<Record<string, SmartAutomation[]>>(this.getSmartAutomationsPath(), (current) => {
-      if (!current || !current[username]) {
-        return current
-      }
-
-      const next = { ...current }
-      next[username] = next[username].filter(x => x.id !== id)
-      return next
-    })
-
-    return { id }
-  }
-
-  private getSmartAutomationsPath() {
-    return join(this.configService.storagePath, 'accessories', 'smart-automations.json')
   }
 
   /**
