+++ src/modules/smart-automations/smart-automations.interfaces.ts
@@ -1,0 +1,8 @@
+export interface SmartAutomation {
+  id: string
+  name: string
+  type: 'smart-light-group'
+  uniqueIds: string[]
+  restoreAfterMs: number
+  enabled: boolean
+}
