+++ ui/src/app/modules/smart-automations/smart-automation-form/smart-automation-form.component.html
