+++ ui/src/app/modules/smart-automations/smart-automation-list/smart-automation-list.component.ts
